Separate filter functionality from builder.rs
