+++ src/builder.rs
@@ -3,24 +3,12 @@ use reqwest::{
     Client, Error, Method, Response,
 };
 
-macro_rules! filter {
-    ( $( $op:ident ),* ) => {
-        $(
-            pub fn $op(mut self, column: &str, param: &str) -> Self {
-                self.queries.push((column.to_string(),
-                                   format!("{}.{}", stringify!($op), param)));
-                self
-            }
-        )*
-    }
-}
-
 #[derive(Default)]
 pub struct Builder {
     method: Method,
     url: String,
     schema: Option<String>,
-    queries: Vec<(String, String)>,
+    pub(crate) queries: Vec<(String, String)>,
     headers: HeaderMap,
     body: Option<String>,
     is_rpc: bool,
@@ -147,19 +135,6 @@ impl Builder {
         self.method = Method::DELETE;
         self.headers
             .append("Prefer", HeaderValue::from_static("return=representation"));
-        self
-    }
-
-    // It's unfortunate that `in` is a keyword, otherwise it'd belong in the
-    // collection of filters below
-    filter!(
-        eq, gt, gte, lt, lte, neq, like, ilike, is, fts, plfts, phfts, wfts, cs, cd, ov, sl, sr,
-        nxr, nxl, adj, not
-    );
-
-    pub fn in_(mut self, column: &str, param: &str) -> Self {
-        self.queries
-            .push((column.to_string(), format!("in.{}", param)));
         self
     }
